Refactor(concurrency): Split the error handler
error_handler.py held four concerns in 450 lines, past the 400-line
reviewability limit the aislop code-quality gate applies.

Split it into a package at the same dotted path: records (the severity
enum and ErrorRecord), handler (the thread-safe
ConcurrentErrorHandler), retry (the with_retry decorator) and
circuit_breaker (CircuitBreaker and CircuitOpenError). The package
__init__ re-exports the previous names, so all import sites are
unchanged.
